sparkscala/requirements.txt
pyspark
kafka-python
pymongo
datetime
pandas
